+++ source/DSCResources/MSFT_ADDomain/MSFT_ADDomain.psm1
@@ -38,7 +38,7 @@ $script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
             Get-ADForest                   | ActiveDirectory
             Assert-Module                  | DscResource.Common
             New-InvalidOperationException  | DscResource.Common
-            Resolve-DomainFQDN             | ActiveDirectoryDsc.Common
+            Resolve-DomainFQDN             | MSFT_ADDomain
             ConvertTo-DeploymentForestMode | ActiveDirectoryDsc.Common
             ConvertTo-DeploymentDomainMode | ActiveDirectoryDsc.Common
 #>
@@ -246,7 +246,7 @@ function Get-TargetResource
         Used Functions:
             Name               | Module
             -------------------|--------------------------
-            Resolve-DomainFQDN | ActiveDirectoryDsc.Common
+            Resolve-DomainFQDN | MSFT_ADDomain
 #>
 function Test-TargetResource
 {
@@ -555,5 +555,46 @@ function Set-TargetResource
         $global:DSCMachineStatus = 1
     }
 } #end function Set-TargetResource
+
+<#
+    .SYNOPSIS
+        Assemble a fully qualified domain name.
+
+    .DESCRIPTION
+        The Resolve-DomainFQN function is used to assemble a fully qualified domain name by appending the domain name
+        to the parent domain name if the parent domain name has been specified. Otherwise the domain name is returned.
+
+    .PARAMETER DomainName
+        The domain name.
+
+    .PARAMETER ParentDomainName
+        The parent domain name.
+#>
+function Resolve-DomainFQDN
+{
+    [CmdletBinding()]
+    [OutputType([System.String])]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $DomainName,
+
+        [Parameter()]
+        [System.String]
+        $ParentDomainName
+    )
+
+    if ($ParentDomainName)
+    {
+        $domainFQDN = '{0}.{1}' -f $DomainName, $ParentDomainName
+    }
+    else
+    {
+        $domainFQDN = $DomainName
+    }
+
+    return $domainFQDN
+}
 
 Export-ModuleMember -Function *-TargetResource